web: move public pages under /home routes
